Fixed pickling of pytorch wrapper.
