+++ bindings/python/tests/test_TorchWrapper.py
@@ -7,6 +7,7 @@ except:
 
 import mpart as mt 
 import numpy as np
+import dill 
 
 if haveTorch:
 
@@ -81,21 +82,6 @@ if haveTorch:
         loss.backward()
         assert tmap2.coeffs.grad is not None
     
-    def test_TorchMethod():
-        opts = mt.MapOptions()
-        tmap = mt.CreateTriangular(dim,dim,3,opts) # Simple third order map
-
-        x = torch.randn(numSamps, dim, dtype=torch.double)
-        tmap2 = tmap.torch(store_coeffs=True)
-        y = tmap2.forward(x)
-        assert np.all(y.detach().numpy() == tmap.Evaluate(x.T.detach().numpy()).T)
-
-        tmap2 = tmap.torch(return_logdet=True, store_coeffs=True)
-        y, logdet = tmap2.forward(x)
-        
-        assert np.all(y.detach().numpy() == tmap.Evaluate(x.T.detach().numpy()).T)
-        assert np.all(logdet.detach().numpy() == tmap.LogDeterminant(x.T.detach().numpy()))
-
     def test_AutogradCoeffAsInput():
 
         opts = mt.MapOptions()
@@ -111,6 +97,36 @@ if haveTorch:
         assert y.shape[1] == dim 
         assert not y.isnan().any()
 
+    def test_TorchMethod():
+        opts = mt.MapOptions()
+        tmap = mt.CreateTriangular(dim,dim,3,opts) # Simple third order map
+
+        x = torch.randn(numSamps, dim, dtype=torch.double)
+        tmap2 = tmap.torch(store_coeffs=True)
+        y = tmap2.forward(x)
+        assert np.all(y.detach().numpy() == tmap.Evaluate(x.T.detach().numpy()).T)
+
+        tmap2 = tmap.torch(return_logdet=True, store_coeffs=True)
+        y, logdet = tmap2.forward(x)
+        
+        assert np.all(y.detach().numpy() == tmap.Evaluate(x.T.detach().numpy()).T)
+        assert np.all(logdet.detach().numpy() == tmap.LogDeterminant(x.T.detach().numpy()))
+
+    def test_TorchPickle():
+        opts = mt.MapOptions()
+        tmap = mt.CreateTriangular(dim,dim,3,opts) # Simple third order map
+
+        x = torch.randn(numSamps, dim, dtype=torch.double)
+        tmap2 = tmap.torch(store_coeffs=True)
+        y = tmap2.forward(x)
+        
+
+        map_bytes = dill.dumps(tmap2, dill.HIGHEST_PROTOCOL)
+        tmap3 = dill.loads(map_bytes)
+
+        y2 = tmap3.forward(x)
+        assert (y2-y).abs().max() < 1e-8
+
 
 if __name__=='__main__':
     test_Evaluation()
@@ -118,4 +134,5 @@ if __name__=='__main__':
     test_Autograd()
     test_AutogradCoeffs()
     test_TorchMethod()
+    test_TorchPickle()
     test_AutogradCoeffAsInput()
